+++ src/undoers/add_image.h
@@ -32,7 +32,7 @@ public:
   AddImage(undo::ObjectsContainer* objects, Stock* stock, int imageIndex);
 
   void dispose() OVERRIDE;
-  int getMemSize() const OVERRIDE { return sizeof(*this); }
+  size_t getMemSize() const OVERRIDE { return sizeof(*this); }
   void revert(undo::ObjectsContainer* objects, undo::UndoersCollector* redoers) OVERRIDE;
 
 private:
