+++ src/undoers/remove_palette.h
@@ -34,7 +34,7 @@ public:
   RemovePalette(undo::ObjectsContainer* objects, Sprite* sprite, int paletteFrame);
 
   void dispose() OVERRIDE;
-  int getMemSize() const OVERRIDE { return sizeof(*this) + getStreamSize(); }
+  size_t getMemSize() const OVERRIDE { return sizeof(*this) + getStreamSize(); }
   void revert(undo::ObjectsContainer* objects, undo::UndoersCollector* redoers) OVERRIDE;
 
 private:
